refactor: keep all the type definitions in a directory
