+++ packages/verifier/client/web-ui/src/types/ClientInfo.ts
@@ -1,0 +1,3 @@
+export type ClientInfo = {
+  status: string;
+};
