+++ packages/verifier/client/web-ui/src/types/Log.ts
@@ -1,0 +1,5 @@
+export type Log = {
+  timestamp: string;
+  level: string;
+  message: string;
+};
