+++ packages/verifier/client/web-ui/src/types/LogFilterOptions.ts
@@ -1,0 +1,8 @@
+import { LogLevel } from "./LogLevel";
+
+export type LogFilterOptions = {
+  page: number;
+  limit: number;
+  filterBy?: LogLevel;
+  search?: string;
+};
