+++ packages/verifier/client/web-ui/src/types/LogLevel.ts
@@ -1,0 +1,7 @@
+export enum LogLevel {
+  all = "",
+  debug = "debug",
+  info = "info",
+  warn = "warn",
+  error = "error",
+}
